Move the meta data to the component instead of with the router
