+++ src/router/routes.ts
@@ -7,24 +7,11 @@
 
 import type { Route } from '@vaadin/router';
 
-import config from '../config.js';
-
-export const pageNotFoundMeta = {
-  title: 'Error: Page not found',
-  description: null,
-  image: null,
-};
-
 export const routes: Route[] = [
   {
     path: '/',
     name: 'home',
     component: 'page-home',
-    meta: {
-      title: config.appName,
-      titleTemplate: null,
-      description: config.appDescription,
-    },
     action: async () => {
       await import('../pages/page-home.js');
     },
@@ -33,10 +20,6 @@ export const routes: Route[] = [
     path: '/about',
     name: 'about',
     component: 'page-about',
-    meta: {
-      title: 'About',
-      description: 'About page description',
-    },
     action: async () => {
       await import('../pages/page-about.js');
     },
@@ -45,7 +28,6 @@ export const routes: Route[] = [
     path: '(.*)',
     name: 'not-found',
     component: 'page-not-found',
-    meta: pageNotFoundMeta,
     action: async () => {
       await import('../pages/page-not-found.js');
     },
